Refactor patient profile adherence display
