+++ src/Basics/Icons/Pill.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import SvgIcon from '@material-ui/core/SvgIcon';
 
-export default function TempIcon({ props, children }) {
+export default function TempIcon({ props }) {
   return (
     <SvgIcon {...props}>
       <svg
@@ -12,7 +12,6 @@ export default function TempIcon({ props, children }) {
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
       >
-        {children}
         <path
           d="M1.77267 22.2273C-0.59089 19.8638 -0.59089 16.0284 1.77267 13.6649L13.6649 1.77267C16.0285 -0.59089 19.8638 -0.59089 22.2273 1.77267C24.5909 4.13623 24.5909 7.97151 22.2273 10.3351L10.3351 22.2273C7.97152 24.5909 4.13623 24.5909 1.77267 22.2273ZM9.14585 8.19447L15.8055 14.8541L21.2759 9.38369C23.1289 7.53074 23.1289 4.577 21.2759 2.72405C19.423 0.871096 16.4692 0.871099 14.6163 2.72405L9.14585 8.19447ZM17.8866 9.8148C17.6786 9.55537 17.706 9.13805 17.9461 8.90801L18.8975 7.95663C19.4264 7.42773 19.7502 6.87216 19.834 6.41064C19.9178 5.94912 19.8385 5.56779 19.3732 5.10249C19.1183 4.85472 19.1151 4.39568 19.3664 4.14434C19.6177 3.893 20.0768 3.89624 20.3245 4.15109C21.0451 4.87169 21.311 5.80017 21.157 6.64846C21.003 7.49675 20.5058 8.251 19.8489 8.90805L18.8976 9.85941C18.6335 10.1351 18.1254 10.1127 17.8867 9.81482L17.8866 9.8148Z"
           fill="black"
